Delete testcontainter.sls
This file was deleted from the files on the master. 
There are already 3 sls files for docker, one to install and enable and two to start up a container, as mentioned in the description for each file. 
There was no need to keep this file anymore. It had a spelling mistake anyway :P